+++ Sources/MapboxDirections/Extensions/Array.swift
@@ -6,3 +6,11 @@ extension Collection {
         return try IndexSet(enumerated().filter { try predicate($0.element) }.map(\.offset))
     }
 }
+
+extension [URLQueryItem] {
+    mutating func override(with params: [URLQueryItem]) {
+        let names = Set(params.map(\.name))
+        removeAll { names.contains($0.name) }
+        append(contentsOf: params)
+    }
+}