+++ Sources/MapboxDirections/Isochrones.swift
@@ -144,7 +144,7 @@ open class Isochrones: @unchecked Sendable {
     /// - Returns: The URL to send the request to.
     open func url(forCalculating options: IsochroneOptions) -> URL {
         var params = options.urlQueryItems
-        params.append(URLQueryItem(name: "access_token", value: credentials.accessToken))
+        params.override(with: [URLQueryItem(name: "access_token", value: credentials.accessToken)])
 
         let unparameterizedURL = URL(path: options.path, host: credentials.host)
         var components = URLComponents(url: unparameterizedURL, resolvingAgainstBaseURL: true)!